+++ core/components/Post.php
@@ -17,11 +17,16 @@ class Post extends Content {
 		/**
 		 * Filter the list of post types not to log for the Posts component. Posts belonging to post types on this list will not be tracked or logged.
 		 *
-		 * @param array $post_types name of the post types not to log
+		 * @param array $post_types names of the post types not to log
 		 *
-		 * @return array array with names of post types not to log.
+		 * @return array names of post types not to log.
 		 */
-		$this->do_not_log = apply_filters( 'coreactivity_post_do_not_log_post_types', array( 'revision', 'attachment', 'shop_order_placehold' ) );
+		$this->do_not_log = apply_filters( 'coreactivity_post_do_not_log_post_types', array(
+			'revision',
+			'attachment',
+			'shop_order_placehold',
+			'user_request',
+		) );
 	}
 
 	public function label() : string {