+++ forms/content-about-whatsnew.php
@@ -27,8 +27,8 @@ include( COREACTIVITY_PATH . 'forms/content-about-minor.php' );
 				<?php } ?>
 
                 <div class="coreactivity-about-counters">
-                    <div><i class="d4p-icon d4p-ui-folder d4p-icon-fw"></i> <strong>26</strong> Components</div>
-                    <div><i class="d4p-icon d4p-ui-radar d4p-icon-fw"></i> <strong>130</strong> Events</div>
+                    <div><i class="d4p-icon d4p-ui-folder d4p-icon-fw"></i> <strong>27</strong> Components</div>
+                    <div><i class="d4p-icon d4p-ui-radar d4p-icon-fw"></i> <strong>140</strong> Events</div>
                     <div><i class="d4p-icon d4p-ui-plug d4p-icon-fw"></i> <strong>12</strong> Plugins</div>
                 </div>
             </div>